Huge performance optimization in Clutter.Color
Unfortunately the way GAPI handles simple structures is incredibly
expensive. Instead of passing the structure by reference for
modification on the native side, it allocates a new instance on
the heap, and passes a pointer to native where it is modified,
and then copied back to managed. This results in massive wastes.

Lots of struct-based APIs here are going to need to be tuned
similarly, which is going to be a giant PITA.