[build] Copy and patch glib-sharp
Okay, this is probably the most insane thing I've done. Basically
we *must* have what is essentially a *one* line patch to glib-sharp.

http://anonsvn.mono-project.com/viewvc/trunk/gtk-sharp/glib/Object.cs?r1=121880&r2=131604

Without this, virtual methods don't work. So this commit does something
batshit insane. It copies glib-sharp from the GAC, compiles the modified
method from C#, and then disassembles the GAC glib-sharp, replacing the
IL of the method that needs patching with the version we just compiled.

With a patched glib-sharp local to clutter-sharp, things magically work.
For now. I hope. WTF.